refactor: MenuGroup package 구조 변경 및 JPA 적용
