--- src/main/java/kitchenpos/dao/MenuGroupDao.java
@@ -1,16 +1,0 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.MenuGroup;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface MenuGroupDao {
-    MenuGroup save(MenuGroup entity);
-
-    Optional<MenuGroup> findById(Long id);
-
-    List<MenuGroup> findAll();
-
-    boolean existsById(Long id);
-}
