+++ src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -1,0 +1,33 @@
+package kitchenpos.menu.application;
+
+import kitchenpos.menu.dao.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.dto.MenuGroupRequest;
+import kitchenpos.menu.dto.MenuGroupResponse;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+public class MenuGroupService {
+    private final MenuGroupRepository menuGroupRepository;
+
+    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+        this.menuGroupRepository = menuGroupRepository;
+    }
+
+    @Transactional
+    public MenuGroupResponse create(final MenuGroupRequest menuGroupRequest) {
+        MenuGroup menuGroup = menuGroupRequest.toMenuGroup();
+        return MenuGroupResponse.of(menuGroupRepository.save(menuGroup));
+    }
+
+    public List<MenuGroupResponse> list() {
+        return menuGroupRepository.findAll().stream()
+            .map(MenuGroupResponse::of)
+            .collect(Collectors.toList());
+    }
+}
