+++ src/main/java/kitchenpos/menu/dao/MenuGroupRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.menu.dao;
+
+import kitchenpos.menu.domain.MenuGroup;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
+}
