+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,6 +1,14 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
 public class MenuGroup {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
