+++ src/main/java/kitchenpos/menu/dto/MenuGroupResponse.java
@@ -1,6 +1,6 @@
-package kitchenpos.dto;
+package kitchenpos.menu.dto;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 public class MenuGroupResponse {
 	private Long id;
