+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -17,7 +17,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import kitchenpos.dao.MenuDao;
-import kitchenpos.dao.MenuGroupDao;
+import kitchenpos.menu.dao.MenuGroupRepository;
 import kitchenpos.dao.MenuProductDao;
 import kitchenpos.dao.ProductDao;
 import kitchenpos.domain.Menu;
@@ -32,7 +32,7 @@ public class MenuServiceTest {
 	@Mock
 	private MenuProductDao menuProductDao;
 	@Mock
-	private MenuGroupDao menuGroupDao;
+	private MenuGroupRepository menuGroupRepository;
 	@Mock
 	private MenuDao menuDao;
 	@InjectMocks
@@ -68,7 +68,7 @@ public class MenuServiceTest {
 		MenuProduct savedMenuProduct = TestDomainConstructor.menuProductWithSeq(NEW_MENU_ID, SAVED_PRODUCT_ID, 1, 1L);
 		MenuProduct savedMenuProduct2 = TestDomainConstructor.menuProductWithSeq(NEW_MENU_ID, SAVED_PRODUCT2_ID, 1, 2L);
 
-		when(menuGroupDao.existsById(SAVED_MENU_GROUP_ID)).thenReturn(true);
+		when(menuGroupRepository.existsById(SAVED_MENU_GROUP_ID)).thenReturn(true);
 		when(productDao.findById(anyLong())).thenReturn(Optional.of(savedProduct), Optional.of(savedProduct2));
 		when(menuDao.save(menu)).thenReturn(savedMenu);
 		when(menuProductDao.save(any())).thenReturn(savedMenuProduct, savedMenuProduct2);
@@ -99,7 +99,7 @@ public class MenuServiceTest {
 	void createNotExistMenuGroup() {
 		//given
 		Menu notExistMenuGroupMenu = TestDomainConstructor.menu("메뉴1", BigDecimal.valueOf(10000), 100L, menuProducts);
-		when(menuGroupDao.existsById(any())).thenReturn(false);
+		when(menuGroupRepository.existsById(any())).thenReturn(false);
 
 		//when-then
 		assertThatThrownBy(() -> menuService.create(notExistMenuGroupMenu))
@@ -111,7 +111,7 @@ public class MenuServiceTest {
 	void createNotExistProduct() {
 		//given
 		Menu notExistProductMenu = TestDomainConstructor.menu("메뉴1", BigDecimal.valueOf(10000), SAVED_MENU_GROUP_ID, Arrays.asList(mock(MenuProduct.class)));
-		when(menuGroupDao.existsById(any())).thenReturn(true);
+		when(menuGroupRepository.existsById(any())).thenReturn(true);
 		when(productDao.findById(any())).thenReturn(Optional.empty());
 
 		//when-then
@@ -124,7 +124,7 @@ public class MenuServiceTest {
 	void createPriceLessThanZero() {
 		//given
 		Menu greaterThanSumOfProductPriceMenu = TestDomainConstructor.menu("메뉴1", BigDecimal.valueOf(100000), SAVED_MENU_GROUP_ID, menuProducts);
-		when(menuGroupDao.existsById(SAVED_MENU_GROUP_ID)).thenReturn(true);
+		when(menuGroupRepository.existsById(SAVED_MENU_GROUP_ID)).thenReturn(true);
 		when(productDao.findById(anyLong())).thenReturn(Optional.of(savedProduct), Optional.of(savedProduct2));
 
 		//when-then
